+++ packages/wolfdesk/test/tickets.spec.ts
@@ -14,6 +14,7 @@ import {
   markTicketResolved,
   openTicket,
   reassignTicket,
+  requestTicketEscalation,
   target,
 } from "./actions.js";
 
@@ -26,7 +27,7 @@ async function findTicket(stream: string) {
   ).at(0);
 }
 
-describe("ticket projection", () => {
+describe("tickets", () => {
   beforeAll(async () => {
     await init_tickets_db();
     await db.delete(tickets).catch((e) => console.error(e));
@@ -37,7 +38,7 @@ describe("ticket projection", () => {
     await dispose()();
   });
 
-  it("should project tickets", async () => {
+  it("projection", async () => {
     const t = target(chance.guid(), "projecting");
     const title = "projecting";
     const message = "opening a new ticket for projection";
@@ -191,6 +192,39 @@ describe("ticket projection", () => {
       expect(snapshot.state.escalateAfter?.getTime()).toBeGreaterThan(
         now.getTime()
       );
+    });
+  });
+
+  describe("reactions", () => {
+    it("should assign agent to new ticket", async () => {
+      const t = target(undefined, "should assign agent");
+      await openTicket(t, "assign agent", "Hello");
+      await app.drain();
+
+      const snapshot = await app.load(Ticket, t.stream);
+      expect(snapshot.state.agentId).toBeDefined();
+    });
+
+    it("should deliver new ticket", async () => {
+      const t = target(undefined, "should deliver new ticket");
+      await openTicket(t, "deliver", "Hello");
+      await addMessage(t, "the body");
+      await app.drain();
+
+      const snapshot = await app.load(Ticket, t.stream);
+      expect(
+        Object.values(snapshot.state.messages).at(-1)?.wasDelivered
+      ).toBeDefined();
+    });
+
+    it("should request escalation", async () => {
+      const t = target(undefined, "should request escalation");
+      await openTicket(t, "request escalation", "Hello");
+      await requestTicketEscalation(t);
+      await app.drain();
+
+      const snapshot = await app.load(Ticket, t.stream);
+      expect(snapshot.state.escalationId).toBeDefined();
     });
   });
 });
